Move non-expression classes out of `expressions` package
