+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/evaluation/values/Thunk.kt
@@ -1,6 +1,6 @@
 package com.github.cubuspl42.sigmaLang.analyzer.evaluation.values
 
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.EvaluationContext
+import com.github.cubuspl42.sigmaLang.analyzer.semantics.EvaluationContext
 
 abstract class Thunk<out ResultType> {
     companion object {
